Rename flow to pipe
